requirements.txt
bootstrap-flask
Flask
sqlalchemy<1.4
Flask-Bootstrap
Flask-Login
Flask-Mail
Flask-Migrate
Flask-Moment
Flask-SQLAlchemy
Flask-WTF
gensim
joblib
jwt
python-dotenv
sqlalchemy-utils
user-agents
Werkzeug
WTForms
email_validator
pandas
mysql-connector-python
mysqlclient